Fixed bug when deleting multiple tags
Tags were removed in ascending order.
For example:
remove(0); remove(1); remove(2); remove(3)
But by the time remove(3) is called the list only contains 2 elements as remove(0) and remove(1) did their job. So remove(3) cannot remove anything.
The fix is to remove elements by descending order so 
remove(3) ; remove(2); remove(1); remove(0)
This works even if the elements are not continuous (1,7,9,10)